copart-deal-finder/requirements.txt
playwright>=1.40.0
beautifulsoup4>=4.12.0
pandas>=2.1.0
numpy>=1.24.0
pyyaml>=6.0.0
requests>=2.31.0
APScheduler>=3.10.0
python-dateutil>=2.8.0
